+++ README.md
@@ -2,9 +2,6 @@
 
 
 -------------------BodyCondition 파일-----------------------                                      
-3 - BodyCondtion은 내부 프로퍼티의 값들이 많이 변하기 때문에 구조체로 선언할 경우 옵셔널 바인딩을 통해
-   값을 변경해줭 하는 것이 번거롭다고 생각해서 클래스 타입으로 선언하였습니다.
-16 - checkCondition 메서드를 통해 클래스 내부 프로퍼티를 출력해주었습니다.
 
 ----------------------Exercise 파일-------------------------       
 3 - Excercise는 처음에 선언을 한 이후에 따로 값 변경이 없을 것이라고 생각해서 구조체로 선언해 주었고
